requirements_ai.txt
# AI packages
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1

transformers==4.42.2
diffusers==0.21.0
accelerate==0.22.0
datasets==2.14.0
huggingface-hub>=0.23.2,<1.0

numpy==1.26.0
pandas==2.1.1
scipy==1.11.0
opencv-python==4.8.1.78
Pillow==10.1.0
tqdm==4.66.1
requests==2.32.1
